chore: code quality cleanup and contract hardening
