--- packages/native/src/ps/types.ts
@@ -1,5 +1,0 @@
-/** Result of a process tree kill operation. */
-export interface KillTreeResult {
-  /** Number of processes successfully killed. */
-  killed: number;
-}
