--- packages/pi-ai/oauth.d.ts
@@ -1,1 +1,0 @@
-export * from "./dist/oauth.js";
