--- packages/pi-ai/oauth.js
@@ -1,1 +1,0 @@
-export * from "./dist/oauth.js";
