+++ packages/pi-ai/src/types.ts
@@ -1,7 +1,3 @@
-import type { AssistantMessageEventStream } from "./utils/event-stream.js";
-
-export type { AssistantMessageEventStream } from "./utils/event-stream.js";
-
 export type KnownApi =
 	| "openai-completions"
 	| "mistral-conversations"
@@ -119,6 +115,12 @@ export interface SimpleStreamOptions extends StreamOptions {
 	reasoning?: ThinkingLevel;
 	/** Custom token budgets for thinking levels (token-based providers only) */
 	thinkingBudgets?: ThinkingBudgets;
+}
+
+export interface AssistantMessageEventStream extends AsyncIterable<AssistantMessageEvent> {
+	push(event: AssistantMessageEvent): void;
+	end(result?: AssistantMessage): void;
+	result(): Promise<AssistantMessage>;
 }
 
 // Generic StreamFunction with typed options
