+++ packages/pi-coding-agent/src/modes/interactive/theme/theme-schema.ts
@@ -1,0 +1,82 @@
+import { Type, type Static } from "@sinclair/typebox";
+
+const ColorValueSchema = Type.Union([
+  Type.String(), // hex "#ff0000", var ref "primary", or empty ""
+  Type.Integer({ minimum: 0, maximum: 255 }), // 256-color index
+]);
+export type ColorValue = Static<typeof ColorValueSchema>;
+
+export const ThemeJsonSchema = Type.Object({
+  $schema: Type.Optional(Type.String()),
+  name: Type.String(),
+  vars: Type.Optional(Type.Record(Type.String(), ColorValueSchema)),
+  colors: Type.Object({
+    // Core UI (10 colors)
+    accent: ColorValueSchema,
+    border: ColorValueSchema,
+    borderAccent: ColorValueSchema,
+    borderMuted: ColorValueSchema,
+    success: ColorValueSchema,
+    error: ColorValueSchema,
+    warning: ColorValueSchema,
+    muted: ColorValueSchema,
+    dim: ColorValueSchema,
+    text: ColorValueSchema,
+    thinkingText: ColorValueSchema,
+    // Backgrounds & Content Text (11 colors)
+    selectedBg: ColorValueSchema,
+    userMessageBg: ColorValueSchema,
+    userMessageText: ColorValueSchema,
+    customMessageBg: ColorValueSchema,
+    customMessageText: ColorValueSchema,
+    customMessageLabel: ColorValueSchema,
+    toolPendingBg: ColorValueSchema,
+    toolSuccessBg: ColorValueSchema,
+    toolErrorBg: ColorValueSchema,
+    toolTitle: ColorValueSchema,
+    toolOutput: ColorValueSchema,
+    // Markdown (10 colors)
+    mdHeading: ColorValueSchema,
+    mdLink: ColorValueSchema,
+    mdLinkUrl: ColorValueSchema,
+    mdCode: ColorValueSchema,
+    mdCodeBlock: ColorValueSchema,
+    mdCodeBlockBorder: ColorValueSchema,
+    mdQuote: ColorValueSchema,
+    mdQuoteBorder: ColorValueSchema,
+    mdHr: ColorValueSchema,
+    mdListBullet: ColorValueSchema,
+    // Tool Diffs (3 colors)
+    toolDiffAdded: ColorValueSchema,
+    toolDiffRemoved: ColorValueSchema,
+    toolDiffContext: ColorValueSchema,
+    // Syntax Highlighting (9 colors)
+    syntaxComment: ColorValueSchema,
+    syntaxKeyword: ColorValueSchema,
+    syntaxFunction: ColorValueSchema,
+    syntaxVariable: ColorValueSchema,
+    syntaxString: ColorValueSchema,
+    syntaxNumber: ColorValueSchema,
+    syntaxType: ColorValueSchema,
+    syntaxOperator: ColorValueSchema,
+    syntaxPunctuation: ColorValueSchema,
+    // Thinking Level Borders (6 colors)
+    thinkingOff: ColorValueSchema,
+    thinkingMinimal: ColorValueSchema,
+    thinkingLow: ColorValueSchema,
+    thinkingMedium: ColorValueSchema,
+    thinkingHigh: ColorValueSchema,
+    thinkingXhigh: ColorValueSchema,
+    // Bash Mode (1 color)
+    bashMode: ColorValueSchema,
+  }),
+  export: Type.Optional(
+    Type.Object({
+      pageBg: Type.Optional(ColorValueSchema),
+      cardBg: Type.Optional(ColorValueSchema),
+      infoBg: Type.Optional(ColorValueSchema),
+    }),
+  ),
+});
+
+export type ThemeJson = Static<typeof ThemeJsonSchema>;
