+++ src/resources/extensions/gsd/types.ts
@@ -2,9 +2,11 @@
 // Types consumed by state derivation, file parsing, and status display.
 // Pure interfaces — no logic, no runtime dependencies.
 
+import type { RiskLevel } from "../../../shared/workspace-types.js";
+
 // ─── Enums & Literal Unions ────────────────────────────────────────────────
 
-export type RiskLevel = "low" | "medium" | "high";
+export type { RiskLevel };
 export type Phase =
   | "pre-planning"
   | "needs-discussion"
