+++ src/resources/extensions/shared/rtk.ts
@@ -1,19 +1,21 @@
 import { spawnSync } from "node:child_process";
 import { existsSync } from "node:fs";
-import { homedir } from "node:os";
-import { delimiter, join } from "node:path";
+import { join } from "node:path";
+import {
+  GSD_RTK_DISABLED_ENV,
+  GSD_RTK_PATH_ENV,
+  RTK_TELEMETRY_DISABLED_ENV,
+  getManagedRtkDir,
+  getPathValue,
+  getRtkBinaryName,
+  isRtkEnabled,
+  resolveSystemRtkPath,
+} from "../../../rtk-shared.js";
 
-const GSD_RTK_PATH_ENV = "GSD_RTK_PATH";
-const GSD_RTK_DISABLED_ENV = "GSD_RTK_DISABLED";
 const GSD_RTK_REWRITE_TIMEOUT_MS_ENV = "GSD_RTK_REWRITE_TIMEOUT_MS";
-const RTK_TELEMETRY_DISABLED_ENV = "RTK_TELEMETRY_DISABLED";
 const RTK_REWRITE_TIMEOUT_MS = 5_000;
 
-function isTruthy(value: string | undefined): boolean {
-  if (!value) return false;
-  const normalized = value.trim().toLowerCase();
-  return normalized === "1" || normalized === "true" || normalized === "yes";
-}
+export { isRtkEnabled };
 
 function getRewriteTimeoutMs(env: NodeJS.ProcessEnv = process.env): number {
   const configured = Number.parseInt(env[GSD_RTK_REWRITE_TIMEOUT_MS_ENV] ?? "", 10);
@@ -23,53 +25,11 @@ function getRewriteTimeoutMs(env: NodeJS.ProcessEnv = process.env): number {
   return RTK_REWRITE_TIMEOUT_MS;
 }
 
-export function isRtkEnabled(env: NodeJS.ProcessEnv = process.env): boolean {
-  return !isTruthy(env[GSD_RTK_DISABLED_ENV]);
-}
-
 export function buildRtkEnv(env: NodeJS.ProcessEnv = process.env): NodeJS.ProcessEnv {
   return {
     ...env,
     [RTK_TELEMETRY_DISABLED_ENV]: "1",
   };
-}
-
-function getManagedRtkDir(env: NodeJS.ProcessEnv = process.env): string {
-  return join(env.GSD_HOME || join(homedir(), ".gsd"), "agent", "bin");
-}
-
-function getRtkBinaryName(platform: NodeJS.Platform = process.platform): string {
-  return platform === "win32" ? "rtk.exe" : "rtk";
-}
-
-function getPathValue(env: NodeJS.ProcessEnv): string | undefined {
-  const pathKey = Object.keys(env).find((key) => key.toLowerCase() === "path");
-  return pathKey ? env[pathKey] : env.PATH;
-}
-
-function resolvePathCandidates(pathValue: string | undefined): string[] {
-  if (!pathValue) return [];
-  return pathValue
-    .split(delimiter)
-    .map((part) => part.trim())
-    .filter(Boolean);
-}
-
-function resolveSystemRtkPath(pathValue: string | undefined, platform: NodeJS.Platform = process.platform): string | null {
-  const candidates = platform === "win32"
-    ? ["rtk.exe", "rtk.cmd", "rtk.bat", "rtk"]
-    : ["rtk"];
-
-  for (const dir of resolvePathCandidates(pathValue)) {
-    for (const candidate of candidates) {
-      const fullPath = join(dir, candidate);
-      if (existsSync(fullPath)) {
-        return fullPath;
-      }
-    }
-  }
-
-  return null;
 }
 
 export interface ResolveRtkBinaryPathOptions {
