+++ src/web/auto-dashboard-types.ts
@@ -1,0 +1,26 @@
+export interface RtkSessionSavings {
+  commands: number;
+  inputTokens: number;
+  outputTokens: number;
+  savedTokens: number;
+  savingsPct: number;
+  totalTimeMs: number;
+  avgTimeMs: number;
+  updatedAt: string;
+}
+
+export interface AutoDashboardData {
+  active: boolean;
+  paused: boolean;
+  stepMode: boolean;
+  startTime: number;
+  elapsed: number;
+  currentUnit: { type: string; id: string; startedAt: number } | null;
+  completedUnits: { type: string; id: string; startedAt: number; finishedAt: number }[];
+  basePath: string;
+  totalCost: number;
+  totalTokens: number;
+  rtkSavings?: RtkSessionSavings | null;
+  /** Whether RTK is enabled via experimental.rtk preference. False when not opted in. */
+  rtkEnabled?: boolean;
+}
