+++ web/lib/gsd-workspace-store.tsx
@@ -66,6 +66,12 @@ import type {
 } from "./session-browser-contract"
 import { authFetch, appendAuthParam } from "./auth"
 import { ContextualTips } from "../../packages/pi-coding-agent/src/core/contextual-tips.ts"
+import type {
+  WorkspaceIndex,
+  WorkspaceScopeTarget,
+  WorkspaceSliceTarget,
+  WorkspaceValidationIssue,
+} from "../../src/shared/workspace-types.ts"
 
 export type WorkspaceStatus = "idle" | "loading" | "ready" | "error" | "unauthenticated"
 export type WorkspaceConnectionState =
@@ -127,29 +133,7 @@ export interface BridgeRuntimeSnapshot {
 }
 
 export type { WorkspaceTaskTarget, RiskLevel, WorkspaceSliceTarget, WorkspaceMilestoneTarget } from "./workspace-types.js"
-
-export interface WorkspaceScopeTarget {
-  scope: string
-  label: string
-  kind: "project" | "milestone" | "slice" | "task"
-}
-
-export interface WorkspaceValidationIssue {
-  message?: string
-  [key: string]: unknown
-}
-
-export interface WorkspaceIndex {
-  milestones: WorkspaceMilestoneTarget[]
-  active: {
-    milestoneId?: string
-    sliceId?: string
-    taskId?: string
-    phase: string
-  }
-  scopes: WorkspaceScopeTarget[]
-  validationIssues: WorkspaceValidationIssue[]
-}
+export type { WorkspaceIndex, WorkspaceScopeTarget, WorkspaceValidationIssue }
 
 export interface RtkSessionSavings {
   commands: number
